+++ main.go
@@ -17,7 +17,7 @@ import (
 )
 
 const (
-	VERSION = "0.2.2"
+	VERSION = "0.2.3"
 )
 
 var configFile string